backend/admin_user/requirements.txt
fastapi==0.116.1
mangum==0.19.0
uvicorn==0.35.0
sqlalchemy==2.0.42
pymysql==1.1.1
python-dotenv==1.1.1
pydantic-settings==2.10.1
email-validator==2.2.0
